Organized Login page into component structure
