+++ frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import './App.css';
-import Login from './Login';
+import Login from './components/Login/Login';
 import Keyboard from './components/Keyboard/Keyboard';
 
 function App() {
