+++ frontend/src/components/Login/Login.stories.js
@@ -1,0 +1,12 @@
+/* eslint-disable */
+import Login from './Login';
+
+export default {
+  title: "Login",
+};
+
+export const Default = () => <Login />;
+
+Default.story = {
+  name: 'default',
+};
