+++ frontend/src/components/Login/Login.styled.js
@@ -1,0 +1,4 @@
+import styled from 'styled-components';
+
+export const LoginWrapper = styled.div`
+`;
