+++ frontend/src/components/Login/Login.test.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom/extend-expect';
+import Login from './Login';
+
+describe('<Login />', () => {
+  test('it should mount', () => {
+    render(<Login />);
+    
+    const login = screen.getByTestId('Login');
+
+    expect(login).toBeInTheDocument();
+  });
+});
